[KAKAOACCNT-2580]Separate routers by role
