+++ src/index.js
@@ -12,9 +12,13 @@ require('dotenv').config();
 const app = new Koa();
 app.keys = [process.env.KOA_APP_KEY];
 const router = new Router();
-const api = require('./api');
+const userApi = require('./api/user.js');
+const authApi = require('./api/auth.js');
+const codeApi = require('./api/code.js');
 
-router.use(api.routes());
+router.use(userApi.routes());
+router.use(authApi.routes());
+router.use(codeApi.routes());
 
 app.use(cors());
 app.use(bodyParser());
